Skip WASM tests on WASM_ENABLED=0
