+++ internal/wasm/sdk/internal/wasm/pool_test.go
@@ -2,6 +2,8 @@
 // Use of this source code is governed by an Apache2
 // license that can be found in the LICENSE file.
 
+// +build opa_wasm
+
 package wasm_test
 
 import (
