+++ internal/wasm/sdk/opa/loader/file/loader_test.go
@@ -2,6 +2,8 @@
 // Use of this source code is governed by an Apache2
 // license that can be found in the LICENSE file.
 
+// +build opa_wasm
+
 package file
 
 import (
