+++ internal/wasm/sdk/opa/loader/http/loader_test.go
@@ -1,6 +1,9 @@
 // Copyright 2020 The OPA Authors.  All rights reserved.
 // Use of this source code is governed by an Apache2
 // license that can be found in the LICENSE file.
+
+// +build opa_wasm
+
 package http
 
 import (
