+++ internal/wasm/sdk/opa/opa_bench_test.go
@@ -1,3 +1,9 @@
+// Copyright 2021 The OPA Authors.  All rights reserved.
+// Use of this source code is governed by an Apache2
+// license that can be found in the LICENSE file.
+
+// +build opa_wasm
+
 package opa_test
 
 import (
